moved tests to test_openai_responses_api.py
